www/qt6-webengine: Update to 6.4.1
Chase the Qt 6.4.1 release as it leads to issues during build if the
version of Qt6 and qt6-webengine is different.

qt6-webengine somehow assumes the current version of Qt6 regardless of
its distfile.  (Also added to the TODO list at the beginning of the
Makefile).

The patches were taken from a more recent Chromium release (commit
96b9a69e7010 in the FreeBSD ports tree) except the newly added patch for
content/common/zygote/sandbox_support_linux.cc.  The latter one is
required to get a successful DEBUG build.