www/qt6-webengine: Fix configure
${WRKSRC}/configure.cmake checks for a binary called 'python3'. This
builds fine in a clean environment prior to this change, but not on a
live system with multiple versions of python 3.x installed. It fails
to configure correctly if the greatest version of python installed does
not also have the html5lib module installed, but, bizarrely, just throws
warnings, not errors, in the comfigure phase and proceeds with a very
brief bunk build and install phase resulting in pkg-plist errors.

a live system with multiple versions of python 3.x installed, it fails.