devel/shiboken6: Install python bytecode
devel/pyside6 was failing to build in poudriere because when it
runs shiboken6, the python bytecode is generated which is a
filesystem violation.

Drop a bunch of unused deps and an unneeded patch.

Add DOCS option. There was a dependency on sphinx, so I assume the
original intention was to install the docs, but they are only built
when the doc target is given.

Formatting cleanups.