x11-toolkits/qt6-graphs: Add port
The Qt Graphs module enables you to visualize data in 3D as bar, scatter,
and surface graphs. It's especially useful for visualizing depth maps
and large quantities of rapidly changing data, such as data received from
multiple sensors. The look and feel of graphs can be customized by using
themes or by adding custom items and labels.

Qt Graphs is built on Qt 6 and Qt Quick 3D to take advantage of hardware
acceleration and Qt Quick 2.

---

Qt Graphs is meant to be the replacement module for Qt Data
Visualization. It uses Qt Quick 3D as the rendering engine to provide
support for native graphics backends, instead of relying on OpenGL 2.1,
like Qt Data Visualization does.

The module will provide most, if not all of the functionality of Qt Data
Visualization. Some advanced features may be incomplete in Technical
Preview.

https://doc-snapshots.qt.io/qt6-6.6/qtgraphs-module.html